+++ src/lib/index.ts
@@ -1,4 +1,10 @@
 import './tailwind/theme.css'
-import { Button, type ButtonProps, type ButtonVariant } from './components'
+import { Button, type ButtonProps, type ButtonVariant, 
+    Input, type InputProps, type InputVariant,
+    Card, type CardProps, type CardVariant
+} from './components'
 
-export { Button, ButtonProps, ButtonVariant }
+export { Button, ButtonProps, ButtonVariant,
+    Input, InputProps, InputVariant,
+    Card, CardProps, CardVariant
+}